Add compatibility with the 'caption' option
Something is wrong with sphinx and the 'container' node (used with the 'caption' option); at some point between the ParsedCodeBlock directive being processed (the run method returning) and the completion of the doctree (the doctree-read event being emitted), the children of the parsed_code_block node get deleted/removed. This shouldn't be anything to do with any transforms, and the children should be set up properly (everything still works when the parsed_code_block is not a part of the container node, i.e. without the caption). Even more oddly, keeping a reference on the parsed_code_block to the deepcopy of the children (without using the deepcopy at all anywhere), the problem is somehow fixed. This 'hack' will be used for now until the reason why this happens is determined.